requirements.txt
Flask==2.3.3
scikit-learn==1.6.1
joblib==1.3.2
psycopg2-binary==2.9.7
numpy==2.0.2
requests==2.31.0
pytest==7.4.3
pytest-mock==3.14.0
